Bug fixes for ALC/CTC, add ITS stack for Mckenzie computation, and update CTC config file
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Branches/v2015.12alc;C64314

+++ development/libraries/DerivedProperties/test/MockGrid.h
@@ -140,11 +140,11 @@ inline double DataModel::MockGrid::maxJ () const {
 }
 
 inline int DataModel::MockGrid::numIGlobal () const {
-   return maxI();
+   return static_cast< int> ( maxI() );
 }
 
 inline int DataModel::MockGrid::numJGlobal () const {
-   return maxJ();
+   return static_cast< int> ( maxJ() );
 }
